Relocate `try_render_mock`
